Add use-mutation
