+++ package.json
@@ -42,7 +42,8 @@
 		"react-dom": "~18.2.0",
 		"react-hook-form": "^7.38.0",
 		"react-i18next": "^12.0.0",
-		"swr": "^1.3.0"
+		"swr": "^1.3.0",
+		"use-mutation": "^2.2.1"
 	},
 	"devDependencies": {
 		"@next/bundle-analyzer": "^13.0.1",
