+++ src/components/RegisterForm/index.tsx
@@ -1,37 +1,21 @@
-import { Button, FormErrorMessage, FormLabel, Input, Text, VStack } from '@chakra-ui/react';
+import { RegisterFields, RegisterFormValues, registerUser } from '@/components/RegisterForm/utils';
+import { Button, FormErrorMessage, FormLabel, Input, Text, useToast, VStack } from '@chakra-ui/react';
 import { useForm } from 'react-hook-form';
-import { post } from '../../utils/network';
-
-enum RegisterFields {
-	EMAIL = 'register-email',
-	FIRST_NAME = 'register-firstname',
-	LAST_NAME = 'register-lastname',
-	PASSWORD = 'register-password',
-}
+import useMutation from 'use-mutation';
 
 export default function RegisterForm() {
-	const { handleSubmit, register } = useForm();
+	const { handleSubmit, register } = useForm<RegisterFormValues>();
+	const toast = useToast();
 
-	function onRegisterAction(formData: Record<RegisterFields, unknown>) {
-		const requestBody = {
-			firstName: formData[RegisterFields.FIRST_NAME],
-			lastName: formData[RegisterFields.LAST_NAME],
-			email: formData[RegisterFields.EMAIL],
-			password: formData[RegisterFields.PASSWORD],
-		};
-		post('/api/auth/register', requestBody)
-			.then((res) => {
-				console.log(res);
-			})
-			.catch((res) => {
-				if (res.errors && res.errors[0] && res.errors[0].message) {
-					console.log(res.errors[0].message);
-				}
-			});
-	}
+	const [onSubmit] = useMutation(registerUser, {
+		onSuccess: console.log,
+		onFailure: ({ error }) => {
+			toast({ title: 'Ops! Something went wrong', status: 'error', description: error?.message });
+		},
+	});
 
 	return (
-		<form onSubmit={handleSubmit(onRegisterAction)}>
+		<form onSubmit={handleSubmit(onSubmit)}>
 			<VStack align="stretch">
 				<FormLabel>Email</FormLabel>
 				<Input type="email" placeholder="Email" {...register(RegisterFields.EMAIL)} />
